Add images verify cmd
This adds the cmd 'images verify', which verifies that previously built
and pushed containerdisks are bootable and that guests are working.

libvmi is used to test containerdisks by spawning VMIs with the
respective containerdisk set as the boot source. After waiting for a VMI
to be running, a defined set of tests is run on the VMI. If all tests
succeed, the value of the containerdisk annotation 'verified' is set to
'true' and the updated manifest is pushed to the registry.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>